Provides complete email validation workflow
