+++ app/controllers/users/omniauth_callbacks_controller.rb
@@ -1,9 +1,13 @@
 class Users::OmniauthCallbacksController < Devise::OmniauthCallbacksController
-  include NYCID::NYCID
+  include NYCID
   def saml
     @user = User.from_omniauth(request.env["omniauth.auth"])
     if @user.email_validated.nil? || @user.try(:email_validated) == false
       email_validation_status = check_email_validation_status(guid: @user.guid)
+      if email_validation_status == false
+        redirect_to validate_email(email_address: @user.email)
+        return
+      end
     end
     set_flash_message :notice, :success, kind: "SAML"
     sign_in_and_redirect @user
